Refactor import routes.
